\name{mi.matrix}
\alias{mi.matrix}
\alias{mi.data.frame}

\title{ 
    Multiply Imputed matrix
}
\description{
    A function that select an imputed data set matrix from a list of imputed data set.
}
\usage{
    mi.matrix( mi.object, m )
    mi.data.frame( mi.object, m )
}
\arguments{
    \item{mi.object}{ mi object containing a multiply imputed data set. 
                        The "mi" object is generated by the mi function.}
    \item{m}{ Index of the imputed data set. The default is 1.}
}
\value{
    A data frame with the missing data imputed.
}
\references{ 
    Andrew Gelman and M. Grazia Pittau,  
    A flexible program for missing-data imputation and model checking, 
    Technical report, Columbia University, New York; Andrew Gelman and Jennifer Hill, 
    Data Analysis Using Regression and Multilevel/Hierarchical Models, 
    Cambridge University Press, 2007.
}

\author{
  Masanao Yajima \email{yajima@stat.columbia.edu},
  M.Grazia Pittau \email{grazia@stat.columbia.edu},
  Andrew Gelman \email{gelman@stat.columbia.edu}
}
\seealso{
    \code{\link{mi}}
}
\examples{
    data(CHAIN)
    imp <- mi (CHAIN)
    imputed.matrix <- mi.matrix (imp) #Imputed matrix from the 1th Imputation (m=NULL, default=1)
    imputed.matrix <- mi.matrix (imp, m=2)  #Imputed matrix from the 2nd Imputation (m=2)
}
\keyword{ models }
